maintner: correctly tombstone "transferred" issues
This work is largely based on jdobry's change (161521)

There are three places where maintner requests an issue (not PR) or
its items from GitHub.

1)  Get issue direct: GitHub returns the transferred issue, we must
    compare repository URLs to detect if it is a transferred issue.
2)  Get issue comments: GitHub returns a 404 or a 301 with Header
    Location = "" for transferred issues.
3)  Get issue events: GitHub returns a 404 or a 301 with header Location
    = "" for transferred issues.

In all three cases we add a mutation to the GitHubIssue marking it
as NotExist, halting further data requests for the issue.

A challenge in this is that the net/http Client treats 301 response
codes with an empty Location header as an error, when this is what
GitHub responds with. For now, check the error returned for "301
response missing Location header"

Change-Id: Iece34e449bd17d8e78ec324cbca5f1d2f4df16d8