maintner: add function to mark an Issue as Tombstoned
The way that maintner queries GitHub for changed issues (and the way
the GitHub API returns Issues) makes it such that if an Issue is
transferred or deleted from a repository, it will never be returned from
the API. The only way to check for transferred issues is to do a scan of
all issues on a Repository and note the differences between the IDs.
This process is not well suited to a syncloop call as it scales with the
number of Issues in the repository.

To combat this, expose a function to mark a given Issue as Tombstoned
(or NotExist). This function can be called by an external process that
performs this more expensive check.

Change-Id: I33caa3608dfb94541665387fb61f1d3e2eda2d95